Add PostgreSQL support and refactor database interactions
