+++ springTest/springTest/src/main/resources/application.properties
@@ -2,5 +2,7 @@ spring.application.name=springTest
 # server.port=8085
 # spring.h2.console.enabled=true
 
-spring.datasource.generate-unique-name=false
-spring.datasource.name=InduwaraDB
+# spring.datasource.generate-unique-name=false
+# spring.datasource.name=InduwaraDB
+
+spring.sql.init.mode=always
