+++ springTest/springTest/src/main/resources/schema.sql
@@ -1,19 +1,10 @@
-CREATE TABLE IF NOT EXISTS Run(
+CREATE TABLE IF NOT EXISTS run(
     id INT NOT NULL,
     title VARCHAR(250) NOT NULL,
     started_on timestamp NOT NULL,
     completed_on timestamp NOT NULL,
     miles INT NOT NULL,
     location VARCHAR(100) NOT NULL,
-    PRIMARY KEY (id)
-);
-
-CREATE TABLE IF NOT EXISTS RunTableForScript(
-    id INT NOT NULL,
-    title VARCHAR(250) NOT NULL,
-    started_on timestamp NOT NULL,
-    completed_on timestamp NOT NULL,
-    miles INT NOT NULL,
-    location VARCHAR(100) NOT NULL,
+    version INT,
     PRIMARY KEY (id)
 );
